requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via pytest
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via pip-tools
django==4.0.3
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
factory-boy==3.2.1
    # via -r requirements.in
faker==13.3.2
    # via factory-boy
flake8==4.0.1
    # via -r requirements.in
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
mccabe==0.6.1
    # via flake8
packaging==21.3
    # via pytest
pep517==0.12.0
    # via pip-tools
pip-tools==6.5.1
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.7
    # via packaging
pytest==7.0.1
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
python-dotenv==0.19.2
    # via -r requirements.in
pytz==2021.3
    # via djangorestframework
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via
    #   pep517
    #   pytest
urllib3==1.26.8
    # via requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
